Support "hic" files as source for interact track

+++ src/main/java/org/broad/igv/hic/ContactRecord.java
@@ -1,6 +1,6 @@
 package org.broad.igv.hic;
 
-public record ContactRecord(int bin1, int bin2, double counts) {
+public record ContactRecord(int bin1, int bin2, float counts) {
     public String getKey() {
         return bin1 + "_" + bin2;
     }
